[RTLIB_CGroups] Setup CGroup enforcing
CGroup enforcing, when required via the proper flag in BBQUE_RTLIB_OPTS, si:
a) enabled: at EXC registration time
b) disabled: at EXC unregistration time

NOTE: for the time being, CGroups enforcing uses the same configuration
for all the EXC of an application. Thus, either:
a) the application uses a single EXC, or
b) the resources assigned should be sufficient to run all the EXCs
The solution a) is preferred.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>