[Tools_OPLB] Update OPList builder to process recipes files
This new version allows to blend OP parameters and metrics within a
recipe file. This should be a convenient support to speed-up exploration
and in general for all these applications where there is a 1:1 mapping
between AWMs and OPs.

Note: by default the tool generate its output on stdout.
However, with this new version, it is possible to define a different
output to be generated by setting the AWM variable OUTF, e.g.

$ bbque-oplb -v OUTF=/tmp/out.cc in.recipe

will generate the output in the /tmp/out.cc source file.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>