[PP_Linux] Add CFS quota enforcing margin and threshold parameters
CFS bandwidth enforcing could be a not necessary overheads in some siutations,
for example when an application is assigned a single CPU with a quota
near the all available bandwidth in this CPU. In this situations the
bandwidth control could be just an unnecessary overhead.

This patch allows to enable the actual enforcing only if the assigned
CFS bandwidht, augmented by a configurable "safety margin", does not
exceed a configurable "threshold value".

The safety margin could be defined using the BBQ configuration option:
  PlatformProxy.CGroups.cfs_bandwidth.margin_pct
which by default is 5%
This margin amount is added to the assigned bandwidth one time for each
PE assigned.

The Threshold value could be defined using the BBQ configuration option:
  PlatformProxy.CGroups.cfs_bandwidth.threshold_pct
which by default is 10%

Signed-off-by: Patrick Bellasi <derkling@gmail.com>