[AWM] Configuration time attribute
Added the "configuration time" attribute. This value should be provided
through the Recipe file, and can be used by an optimization policy.
As well as for the Value, the configuration time features a pair of values: the
absolute one provided in the recipe, and the normalized one (considering
the times of the AWMs in the same recipe).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>