[Recipe] AWMs attributes normalization
A new attribute has been added to the WorkingMode class: the configuration
time. This attribute is managed just like the Value, being therefore
normalized on the basis of all the configuration times of the AWMs included
in the same Recipe.
This addition has led to a refactoring of the Recipe class, for the part
concerning the "normalization" support data structures and member functions.
However it is important to take into account that the Configuration time
is normalized in a different way with respect to the Value. This in order
to have a more uniform distribution of values in the range [0, 1].

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>