[SCReconfig] Configuration time based computation
If the configuration time has been profiled and provided with the Recipe,
the Reconfiguration index is computed as the difference between 1 and
the normalized time. This means that the AWM featuring the highest
configuration time will return an index equal to 0, while the AWM with the
lowest configuration time will retunr an index equal to 1.

Whether the configuration time has not been profiled, the contribution will
return an index computed as done in the previous version, moved into the
ComputeResourceProportional() function.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>