[BuildSystem] Refactor CMake common options
The set of CMake common options defined by the BOSP configuration
are not collected within the single BBQUE_CMAKE_OPTS local
variable before to configure CMake.

Moreover, the new BOSP building system, does not force anymore the
building folder to be two levels up. Thus, now we define explicitly
"../.." as the folder for sources to be build.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>