[RA] Resource path prefix
The idea is that the "prefix path" should somehow reference a "computation
node" of a system, e.g., on a NUMA machine the union of a CPU and its own
memory node.
By keeping track of a "prefix" resource path is it possible to add
flexibility in the validation of recipes, and simplify some resource
binding operations.
By the Application Working Mode side it is expected that a Recipe could
feature resource paths of type ".cpu.pe". Thus, its app to the AWM
validation to add the "prefix path" such that the Scheduling Policy could
work on valid and complete resource paths.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>