[RA] GetAmountFromUsagesMap(): changed resource path check
To retrieve the information under a given "scope" resource path, the
ID-based check for the existence of the scope resource path has been
replaced by a check based on retrieving a pointer to the ResourceIdentifier,
which works much better. This because it is possible to have resource
usages map featuring paths of mixed/template type.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>