[RI] DONT_CARE case in comparing objects
Added the 'DONT_CARE' return code in the Compare() member function.
This is the case of when at least one of the two ResourceIdentifier objects
is of type UNDEFINED (string: "*").

Example:  "cpu1".Compare("*") = DONT_CARE

This feature is expected to be the basis for resource query featuring path
like "sys0.*.pe". For instances, this would return all the PROC_ELEMENT
resources, without distinguish the parent resource type (CPU, GPU, ...).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>