[RA] Alignment to the new state machine
The synchronization, the resource booking and update have been subjected
to refactoring in order to manage the RA working according to the new
state machine.

More in detail, a single mutex is used. This replace the usage of a mutex
for the synchronization state only. Furthermore the boolean data member
'sync_ssn.started' has been removed, since the SYNC state is now checked
through the data member 'status'.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>