[PP_Linux] Remove logging of percent sign
For the time being, the Log4CppLogger class has some issue on formatting
log messages which contains a "%" sign. This is probably due to the parsing
of these parameters via va_start/va_end.

This patch simply removes the dumping us such a sing, which is not required
in this specific case.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>