[Logger] Update color formatting code
When cross-compiling Log4CPP the "logger.log" on the PatternAppender seems
not to properly manage va_list of arguments.
Only the first string is produced in output, which is unfortunately just the
color formatting escape string, while the actual message is lost.

This patch update the color formatting code by adding the required support
to completely format and colorize the log messages within the Log4CppLogger
class methods. The properly formatted string is than passed to Log4CPP
for the final dispatching on the required output appender.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>