[RTLIB_RPC] Disable GGap while the EXC is in synchronization
This patch allows to send a GGap to the BarbequeRTRM only if the EXC is
not already in synchronization mode.
This fixes a critical problem since, if the GGap request overlaps with one
of the reconfiguration messages, the responses to this message could be
lost (timeout) and thus BBQ disable the application.

By the way, it makes sense to disable the GoalGap interface while an
EXC is already undergoing a reconfiguration.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>