[RTLIB_RPC] Add support for GGap filtering
Once an EXC asserts a GoalGap, the RTLib could decide to not forward that
request if it is not greater than a predefined and configured value.
This path add the required support to define the default value for that
filtering threshold.

A Kconfig option:
Barbeque -> Run-Time Library -> GoalGap Forward Threshold
is expose to define the default threshold value.

Moreover, a BBQUE_RTLIB_OPTS flag esposed to define a run-time value for
that threshold, for example:
   BBQUE_RTLIB_OPTS="t45"
configures a 45% threshold value for GoalGap forwarding.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>