[AM] Add support for GGap threshold optimization
Once an EXC asserts a GoalGap, an optimization event is generated within
the BarbequeRTRM only if the asserted value is greater than a predefined
and configured value.
This path add the required support to define the default value for that
filtering threshold.

A Kconfig option:
  Barbeque -> Advanced Options -> GoalGap Optimize Threshold
is expose to define the default threshold value.

Moreover, a configuration file option:
  ApplicationManager.ggap.threshold_optimize
could be used to tune that threshold.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>