[Kconfig] Memory locality selection
This enables the selection of the memory locality to consider in the
resource registration and the resource allocation policy.
According to the terminology used on Linux systems, if the "L3 cache" is
selected, the BarbequeRTRM should register resource of type "Socket" as CPU
resources. In the case of "L2 cache" instead, the CPU resources will be
the "Cores".

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>