[Config] Fix deployed bbqued path
In case of deployment on systems different from the "native", like the case
of cross-compilation for an ARM board, the root directory considered to
search the deamon script was the value of CMAKE_INSTALL_PREFIX, that
commonly points a directory related to the host machine.
In this patch this behaviour has been fixed by using
CONFIG_BOSP_RUNTIME_PATH instead.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>