[AWM] UBD: Multi-BD resource binding for PE/MEM resource types
This fix allows to generate resource bit-masks for PEs/MEMs when dealing
with multiple Binding Domains at once (i.e. scheduling an application on
a multi-BD resources partition).

Usually, the Resource Binding is performed by selecting all the desired
resources from a `resource pool`, which consists of a sub-tree of the
system resources (e.g. the content on a specific Binding Domain).
The `resource pool` is identified by its resource path
-- e.g. sys0.cpu1.pe, where `cpu1` represents for instance the binding
domain selected by the scheduling policy (in this case the resource pool
comprehends all the PEs owned by the binding domain (i.e. CPU) 1).

However, in case of multiple binding domains, the resource path
pertaining the selected resources is generic (e.g. sys.cpu*.pe).
Consequently, the bit-mask representing the selected resources cannot be
created, because the binder cannot identify a resource pool.

The issue arises only in the case of PEs and MEMs binding, because in the
case of multi-BD scheduling these are the types of resources that can
belong to different BDs, yet being allocated to the same application.

To tackle this issue, in the case PEs and MEMs the resource pool is
manually defined. The pool contains the union of all the PEs/MEMs of the
system, i.e. the union of the resources owned by a CPU with any ID. The
GetMask method has been also modified to allow the usage of generic
scopes.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>