[AWM/APP/RA] AWM resource binding update
With this patch, the bitsets (or bitmaps) related to the resource binding
performed by the scheduling policy are updated after a successful
synchronized resource acquisition (performed by the Resource Accounter).
Previously this was performed just after the ScheduleRequest() called onto
the Application.

This enables the possibility of providing to the PlatformProxy all the
bitsets related to the resource binding made by the scheduling policy in
a simpler and more readable way: using the WorkingMode member function
BindingSet() with the resource type specified as argument.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>