[AWM] Fix misbehaviour in resource usage addition
When AddResourceUsage() is called from the Recipe Loader, in case of
missing resource, the ResourceUsage object is now added as well to the AWM.
This allows to "skip" the AWM through the validation step
(calling Validate()).

This fixes a previous misbehaviour, where for example an AWM included a CPU
and a GPU resource usage. Whether the platform did not feature a GPU, the
only CPU usage was added, and in the validation step everything went fine
without invalidating the AWM. The worst effect was that such AWM could
have the highest value, but reasonably with a lower amount of CPU resources
compared to another one. Missing the GPU, the application would have been
served with the highest value for a resource set not consistent to what
was expected.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>