[SchedPol_YaMS] Fix multiple binding domains evaluation
It has been fixed a bug on heterogeneous systems (CPU+GPU) according to
an AWM requiring only the CPU was not selected anymore. Once evaluated the
CPU bindings, and passed to the GPU, in case of AWM not requiring a GPU
processing quota the code did not enter the block featuring the addition
of the scheduling entity into the ordered list.

Moreover, with this patch the amount of processing resource required is
evaluated before looping amount the binding options. In case of '0', the
loop is skipped by returning immediately.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>