[PM_CPU] PEs/Cores mapping initialization
Introduced the `core_ids` map structure, where core_ids[PE_ID]= Core_ID.
In this way, when dealing with a PE, obtaining the ID of the core which
owns the PE becomes straightforward. This can serve, for instance, when
dealing with hyper-threaded devices (e.g. each core contains multiple) PEs.
The map is initialized in the CPUPowerManager constructor.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>
Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>