[Resource] Power profile information
It is now possible to collect data about power/thermal status of the
resource.

This can be enabled with EnablePowerProfile() by specifying the number of
samples to collect, for each profile information required (e.g. LOAD,
TEMPERATURE, FREQUENCY, etc...), needed to compute an exponential mean of
the value. If no arguments are provided, the default configuration is
loaded (BBQUE_PM_DEFAULT_SAMPLES_WINSIZE).

The data can be update by calling the suitable member function of the Power
Manager and passing the returned value to the UpdatePowerInfo() function,
specifying the type of profile information to update.

The member function GetPowerInfo() is provided to return the value of a
power profile information. It is also possible to require the latest
sampled value (ValueType::INSTANT) or the computed mean (ValueType::MEAN).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>