[PP] Power monitoring
The Platform Proxy worker thread now can perform the monitoring of the
power/thermal status of the hardware resources.

To enable this feature, a suitable power management policy should register
the set of resources to monitor with PowerMonitorRegister() and then
call PowerMonitorStart(). The member function PowerMonitorSample() is in
charge of collecting the status information required about the registered
resources. PowerMonitorStop() stops the monitoring activity.

Afterwards, the policy can retrieve the information collected from the
Resource descriptors.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>