[RPC] Runtime profiling messaging functions
The profiling is actually available only for OpenCL applications.
Applications must be launched with the BBQUE_RTLIB_OPTS variable properly
set to "o".

The BbqueRPC class collects profiling statistics at runtime. When required
from the resource manager, it returns through the RPC Fifo client member
functions, the packets with the profiled data. The type of data provided
so far are:
- KERNEL EXECUTION TIME: The cumulated amount of time spent in executing
  the kernel on the device.
- MEMORY TRANSFERS TIME: The cumulated amount of time spent in executing
  data copy and memory transfers commands.

The data provided are the difference between the current amount of
cumulated time, and the cumulated time samples during the previous request.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>