[SchedPol_CLoves] First version
The policy focuses OpenCL workloads on heterogeneous systems.
To properly work it requires that:
  - The periodic mode, i.e. that the ResourceManager.opt_interval option
    in the configuration file could be set
  - The power monitoring is enabled (both CPU and GPU), such that is could
    be possible to get the current load level of the devices
  - The OpenCL application profiling is enabled (BBQUE_RTLIB_OPTS="o"), to
    provide runtime statistics about time spent to execute on the device,
    and time spent for memory data transfers.

The current version works as follows:
1) Profiling information are stored at AWM level, such that the policy
could evaluated the best AWM, considering the Value and the runtime
profiling. If an application provides and AWM for the execution of kernels
on GPU, and another for the execution on CPU, this means to estimate the
"best" device type for executing the application kernels.
2) Given the device type chosen, if there are multiple devices available,
the one with the lowest load level is picked
3) The application is enqueued for the execution on the device into a
priority queue. The priority is a metrics that takes into account how much
the application has been served by the assignment of the AWM selected. In
other words, higher priority is given to application that used the device
for a less amount of time, weighting this time value with the AWM value.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>