[SchedPol_CLoves] Device queue balancing based on length
Actually, selecting the device queue according to the load is a not trivial
task, since it requires also to know the load impact produced by the
allocation of the kernel, in order to properly perform balancing.
Give the task complexity, the selection of the queue is simply performed by
considering the length of the queues. In other words, the shortest queue is
selected.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>