[PIL] New platform description file naming
The platform description file (.bpl) name is now subject to a new convention:
If the "hardware identifier string" is defined (CONFIG_TARGET_ID),
the file takes the name ${CONFIG_TARGET_ID}.bpl

If the file does not exist, the init script will look for `hostname`.bpl,
like in the past. If the check fails, the PIL generator comes in the play.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>