[SC_Value] Negative Goal-gap management
The scheduling contribution has been deeply rewritten.
The old "nap weight" is (actually) unused. Basically, the standard behavior
is to return the static Value specified in the recipe, unless a Goal-gap
has been received.

If the Goal-gap is positive, an AWM with value higher than than the current
one should be selected. Conversely, a negative Goal-gap leads to look for
AWM with lower values. The best candidate is expected to be the AWM with the
value closest to the "ideal value", which is computed starting from the
value of the current AWM and adding (subtracting) the Goal-gap value.

The contribution value will result equal to 1 minus the (absolute)
difference between candidate AWM value and "ideal" value.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>