[SchedPol_Tempura] Baseline implementation
The basic idea of the policy is that starting from thermal and/or energy
constraints, the amount of resource (budget) to allocate is defined
accordingly.

Given all the applications and their priorities, the number of "slots" is
computed. In this way, the resource budget is converted into slots, such
that the policy could assign the amount of resources proportional to the
application priority.

This is the first policy that builds an AWM online instead of iterating
over the list coming from the Recipe.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>