[APP] History-based goal-gap
Very first approach to history-based goal-gap.
If the goal-gap is still positive, the policy has not properly
served the application. In this case add to the previous
value the current goal-goap received.

This has a double effect: 1) in case of positive value
history and positive current value this increase the entity
of goal-gap seen by the policy, pushing it to satisfy the
application. 2) in case of positive value history and negative
current value, the previous request is mitigated since it
now results outdated.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>