[RTLIB] SetCPSGoal function
Added a function to allows the application developer to set a cycle rate
goal to the specified Cycles-Per-Second (CPS) value.
The RTLIB will transparently monitor the performance "Goal-Gap", and in
case of gap absolute value over passing a configurable threshold, the
SetGoalGap() is automatically called.

This relieves the developer from the burden of computing the "Goal-Gap" by
itself and explicitly calling the SetGoalGap(), reducing the amount of
code to place in the onMonitor member function.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>