[RPC] Goal-Gap forward rate exploitation
Implemented a sort of filter frequency, at the granularity of execution
cycles. Accordingly, this is a further filter applied to the Goal-Gap
assertions. It acts ahead of the amplitude filter (threshold).

Specifically, the filter consists of checking if the "enough" cycles are
passed since the last assertion. Such number of cycles is the value of the
"Goal-gap Forward Rate" parameter.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>