[Linux_PP] Fix resource mapping
To extract the resource mapping from a UsagesMap during the resource
allocation/enforcement step of the Platform Proxy, it is mandatory to
provide the reference to the application and the resource view.
Without this pair of information the GetMask erroneously returned the
entire set of resources included under a given scope (binding domain /
node).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>