[RB] Fix deep resource binding
In case of resource binding on processing elements (PROC_ELEMENT) or memory
(MEMORY), when the Platform Proxy tried to retrieve the set of resources
under a specific node (CPU), the GetMask function returned an empty mask.

This misbehavior has been fixed by accessing to the binding resources list
also in case of "scope identifier" found equal to ANY or NONE, and by
properly rearranging the comparison statements.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>