[SchedPol_Tempura] First energy budgeting support
The energy budgeting is performed only if the battery management support
has been enabled at compile time. In such a case, if the resource power
budget computed from the system power budget is positive, the power budget
returned for the computation of the resource budget is the minimum
between the values computed for the thermal constrain and the energy
budget.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>