[Linux_PP] Fix multi-node resource binding
So far the attributes of the RLinuxBinding structure were overwritten after
each iteration of the loop over the binding nodes. This misbehavior has
been fixed by acting in the GetResourceMapping, in such a way that the
function append into the attributes the information retrieved for each node,
instead of ovewriting it.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>