[RecipeLoader_RXML] Fix HW platform parsing
The plugin worked incorrectly during the platform matching stage.
In case of HW target specified (<platform hw="..."), the attribute were
not properly taken into account.

The overall behavior has been reviewed such that, if the matching fails in
searching the required HW section, the section node returned is the last
matching the platform ID (e.g., 'org.linux.cgroup').
If also the ID does not match, the node returned is the one with the id
attribute set to 'generic'. If this has not been specified in the recipe,
the node returned is the first one parsed from the recipe.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>