[SchedPol_YaMS] Fix resource binding evaluation
The ResourceAccounter function GetUsageAmount refactoring introduced a
bug in YaMS scheduling policy, since the new function definition requires
a resource state view to return the correct resource usage value.

YaMS evaluates the resource requests performing the query on the resource
usages map coming from the recipe parsing, hence not bound to any platform
resource. The results was an always 0 returned value.

This problem has been fixed by introducing a new version of the function.
Note that this version does not need the application and the resource state
view references.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>