[RPC/RTLIB] Signed AWM identification number
Since new BarbequeRTRM policies can work also without considering the
application recipe, the AWM identifier returned to the application may have
no sense. Therefore, this commit introduces the possibility of sending
a negative value (e.g., -1) to notify that the AWM is not related to none
of the ones included in the recipe.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>