[BospShell] Fix damaging cleanup
In case of double 'bbque-startd' the result was that the daemon was not
restarted, but the applications were not able any more to communicate with
the daemon. The source of the problem was that in the command defined in
the 'bosp_init.env' it was performed an erroneous 'bosp-cleanup', wiping
out the stuff under "out/var/..." before checking if the daemon is already
started.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>