[RU] Resources list filling policy
This commit adds the possibility of specifying a filling policy for the
resources list. This approach is alternative to the per-resource
fine-grained binding, that a resource allocation policy can perform.

With the filling policy, the Resource Accounter, can take care of this
task by using a predefined schema, specified through the enumerated class
Usage::Policy. The policy defines how the amount of resource required must
be spread over the set of physical resources that the Usage object is
referencing.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>