[Kconfig] Max number of resources option
This has been made configurable, replacing the MAX_R_ID_NUM macro
statically defined in res/identifier.h (value=64), with the
BBQUE_MAX_R_ID_NUM that is set to the value CONFIG_BBQUE_RESOURCES_MAX_NUM
- 1.

The default value of such an option is set to 16. Reasonably this change
should impact positively the memory occupancy of the framework.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>