[RTLIB/RPC] Number of processing elements assigned
PreChange message extended to include the number of processing elements
assigned. This information is required since the scheduling policy can
allocate a given amount of processing quota, spreading it on a variable
amount of processing elements (cores).

For instance a processing quota of 100% can be allocated on a single cores
as well as being distributed over multiple cores.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>