[AM] Synchronized DisableEXC()
Before performing the cleanup of an Application (EXC), the call wait the
Resource Accounter synchronization phase to complete.

This solves a bug experienced when the latency of the BarbequeRTRM was
higher than the very short completion time of the application. In such a
case the completion of the application triggered a termination event that
was handled by the Application Manager (though release-agent and command
interface), which used to call DisableEXC while the ResourceAccounter
was still performing the synchronization step.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>