[SchedPol_Tempura] Little CPU fixed budget
Since, it is currently not possible to constrain an application by
specifying two CPU bandwidth quota constraint for two separate CPU set
(e.g., Little CPU set and Big CPU set), the policy will use the approach
of setting a fixed amount of CPU quota to assign (virtually) to the
Littel CPU set. Accordingly, when the application does not generate a load
such that its threads are migrated into the Big set, it is still able
to use the Little set.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>