[PP_PL] Linux process listener module
A new module (ProcessListener) has been added to the platform proxy that
is capable of monitoring other running processes in the Linux through
the process connector API. At the time being, the class is instantiated
by the platform_proxy as a worker and it will log the signals (EXEC,
EXIT,...) of other processes as error events.

To-Do:
1) Take the proper action on all the events (for now only exec and exit
are reported).
2) The log messages should be Debug messages. At the time being the
messages are of the type Error as the configuration didn't let it to
produce Debug messages.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>
Signed-off-by: Sasan Golchin <ahmad.golchin@mail.polimi.it>