[RM] Triggering optimization even if there are no apps to handle
When an application terminates, the RM triggers a re-schedule.
Usually, the re-schedule is not triggered if there are no running
or waiting applications; however, optimization should be triggered
even if there are not applications to run: in fact, the scheduler
may want to put the system in a low-power state or to reset some
learning data, etc.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>