[Plugins] AgentProxy introduction
This plugin must provide connectivity to the BarbequeRTRM such that it
becomes possible to deploy the framework on distributed platforms. The
goal is to enable the resource manager to operate in a multi-agent based
configuration. This commit introduces the first definition of the plugin
interface.

An instance of BarbequeRTRM should be able to retrieve data from another
instance (running on a different system node) about status of the resources, the
current workload and the communication channel.

Moreover, a further idea is that more instances (system nodes) can join or
not the same partition. Accordingly a distributed system can be partitioned
such that different subsets of nodes can be managed separately.

Different mode will be defined for the distributed configuration. The
simplest one is a hierarchical centralized approach, where the resource
allocation policy is ran by a single (master) instance of the resource
manager. As a consequence the AgentProxy must provide an interface to
send a schedule request to a remote node.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>