[GRPC_AgentProxy] Initial version
This AgentProxy plugin is base on the gRPC library by Google Inc.

The plugin creates an instance of AgentServer which launches a thread to
handle all the incoming requests.

The server logic is mostly implemented in the AgentImpl class which is
responsible of filling the reply messages.

The AgentClient objects instead are used to manage the outcoming requests,
i.e., to communicate with the other BarbequeRTRM instances.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>