[BDM] BindingManager early introduction
This is a new component aiming at factorizing the logic related to the
resource binding process, common to all the policies. As it already happens
for other managers, this is expected to be a container of several resource
binding policies. Then, it will be up to the specific scheduling policy to
choose one of the binding policies already provided by the BindingManager,
or to implement a new one.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>