[RU] Refactoring introducing bitset masks
The resources requested/assigned are now tracked not only through the
resources list, but also exploiting ResourceBitset type masks.

The bit set in the mask represents the id numbers of the resources in the
list. For example, for the resource path "sys.cpu.pe" the mask will include
the id number of the processing elements assigned.

Accordingly the iterators have been removed.

This change has been introduced to speed up and simplify some comparison
operations and the resource binding process.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>