[RB] Bind function review and overloading
Return void type instead of size_t. The caller can check the size of the
'out_map' object to get the same information.

The overload version introduced allows the caller to perform resource
binding giving the source resource assignments map, a resource path and a
mask (ResourceBitset) with the set bits representing the id numbers of the
resources to pick, given the list referenced by the resource path.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>