[WM] Assign id number to SYSTEM in prefix path
When a resource request is added to a working mode object, and the prefix
path is retrieved, it has been added a default behaviour according to
which, the SYSTEM level of the prefix path must have id=0.

This is needed to simplify the resource binding when the SYSTEM part of
the resource path is not affected by the binding performed by scheduling
policy.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>