[WM] BindResource refactoring and overloading
* A new BindResource() version has been introduce to allows the policy to
specificy a resource path and a ResourceBitset object to directly map the
path to a set of resource descriptors. The ResourceBitset set bits refer to
the id numbers of the descriptors to bind to.

* The reference numbering of the binding maps has been changed. The map
has been replaced by a vector. Therefore we do not need anymore of the
hash function to generate a key. This change has been introduce to make
possible the use of both the version of BindResource(), without struggling
to retrieve the on going binding map.

* The function BindingStr() became useless and it has been removed.

* The vector of scheduling binding maps is cleared once the synchronization
is performed (SetResourceBinding() call).

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>