[AP] Fix assignment when OpenCL is enabled
In case of OpenCL support enabled, the missing assignment of a GPU leads
to retrieve the resource id number of the CPU device. However, if the
OpenCL runtime does not access devices of type CPU, the pointer to list
(of id numbers) returned is null. This condition was not managed in the
past, leading to easily experiencing segmentation faults.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>