[BAT] Fix charge monitoring for v4.x kernels
It is still not completely clear, but the sysfs battery status interface
may exhibit a different layout depending on the machine.

For instance  some machines may have "charge_*" attribute files to export
data about the charge, while other may have "energy_*". This is now checked
in order adapt the BarbequeRTRM to the current case.

Similarly, to get the current power supplied we could find "power_now" to
directly provide us this data, or we could have "current_now" to read the
current discharging rate, and thus computing the power considering the
data read in "voltage_now" also.

The class has been modified to handle these cases.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>