[APP-RPC-AP-LPP] CGroup distributed actuation
Moved control groups actuation from the Platform Proxy
to the RPC. The global manager, i.e. the BarbequeRTRM,
assigns a resource budget (CPU for now) to application,
specifying which PEs are shared and which ones can
be used in isolation. It's up to the local manager (RTLIB)
to choose how to use that budget. Too high/low budgets are
notified to the global manager in advance, using a statistical
analysis to understand when to forward runtime profiles.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>