[SchedulerPolicyIF] GetSlots member function introduced
The function has been implemented for the first time in the TEMPURA policy.
It provides a way to allocate an amount of resources proportional to the
application priority. Since this function can be useful for several policies
it has been now defined and implemented in the father class.

The "slots" are computed by summing up the "reverted" values of the
priorities of all the applications. A "reverted priority value" is obtained
for example as follows:
- MIN PRIO VALUE = 5,
- MAX PRIO VALUE = 0
- NUM SLOTS for application with PRIO 0 = (MIN - PRIO) + 1 = 6
In other terms a PRIO 0 application should receive 6 slots of resource,
while a MIN PRIO will receive 1 slot.
The total amount of resource can be divided by the number of slots returned
by the function, in order to have a resource amount unit to multiply for.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>