[SchedulerPolicyIF] ForEachReadyAndRunningDo member function
Looping over READY and RUNNING (i.e., ACTIVE) application is a common
activity in most of the policies. Therefore this function has been
introduced to avoid code replication and simplify the policy implementation.
With a single line of code it is now possible to specify a function
which performs this iterations and calls a given function taking a pointer
to Application as argument.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>