[CPU_PM] Thermal monitoring configuration
The focus has been moved onto the configuration file.
This because the sysfs path for the thermal sensors changes among kernels
version and because the system may include multiple sockets, thus exposing
multiple directories including thermal monitoring stuff.

Therefore the user can specify in the configuration file the number of
sockets and the path for each of them.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>