[WM] Multi-threaded monitoring
The power monitoring activity is now performed on a multi-threaded fashion.
The number of threads is specified in the configuration file.
Each thread will monitor a subset of resources in parallel.
In case the number of resources is lower than the number of threads
specified, this one is automatically set to 1.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>